Rename 'Git' module to 'Git_Windows' for sanity reasons
