+++ classes/__init__.py
@@ -4,4 +4,4 @@ import platform
 from ._Post import Post, PostParseError
 if 'windows' in platform.platform().lower():
     # Only make our Git module available if we're on Windows.
-    from ._Git import Git
+    from ._Git_Windows import Git
